[fix] gccの警告に従い修正
以下の細々とした警告に対応:
 - switch-case文でbreakが無い（その前の関数の中でプログラムが終了するが、一応付加しておく）
 - 他ファイルに移した未使用のstatic関数
 - 変数の型の誤り
 - 符号付き変数と符号無し変数の比較
 - strlenにNULLが渡される可能性
 - 未使用変数の存在
 - 構造体の初期化メンバの不足
 - for と同じ行に2つ以上の文がある

+++ src/combat/melee1.c
@@ -569,7 +569,7 @@ static MULTIPLY mult_slaying(player_type *player_ptr, MULTIPLY mult, const BIT_F
 	};
 
 	monster_race* r_ptr = &r_info[m_ptr->r_idx];
-	for (int i = 0; i < sizeof(slay_table) / sizeof(slay_table[0]); ++i)
+	for (size_t i = 0; i < sizeof(slay_table) / sizeof(slay_table[0]); ++i)
 	{
 		const struct slay_table_t* p = &slay_table[i];
 
@@ -612,7 +612,7 @@ static MULTIPLY mult_brand(player_type *player_ptr, MULTIPLY mult, const BIT_FLA
 	};
 
 	monster_race* r_ptr = &r_info[m_ptr->r_idx];
-	for (int i = 0; i < sizeof(brand_table) / sizeof(brand_table[0]); ++i)
+	for (size_t i = 0; i < sizeof(brand_table) / sizeof(brand_table[0]); ++i)
 	{
 		const struct brand_table_t* p = &brand_table[i];
